pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.5",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-supervisor>=0.0.29",
    "langsmith>=0.4.14",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.1",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
]
